Revised sdk interface
This module is used to interface with openstacksdk.  The main problem we
encountered is that the default list() operation carries DB session.
There are cases where our list generation is from in-memory data.  To
overcome that we provide a list_short() method to Resource class.

The SenlinConnection class is not useful at the moment.

The UserPreferences is a standard way to specify desired API version,
region name, an API visibility (public, private, admin). We provide a
subclass here to deal with the respective requirements.