Make event-list sort_keys all workable
Currently when listing events, user can sort events by keys like
'user' and 'obj_name', but these columns are not in the list which
makes the sorting meanless.
Because 'user' is almost always keeping the same, so don't use it
as a sort key, adding 'level' as a sort key.

Change-Id: I3b298f660faf81e69431320248bc6b342faff3b2
Closes-bug: #1528103