Remove cluster-policy-enable/disable
The cluster-policy-enable and cluster-policy-disable APIs are now
useless since we have removed almost all other properties from a
cluster-policy binding. The enable/disable logic has better be done
through the cluster-policy-update operation which is more generic than
these two. So we have deleted them from senlin-api and sdk.

This patch adapts the client to this change.

Change-Id: I3c3a58ab7e75faec21400bbbad0b5e8420790d51