Handle ResourceNotFound when resource is not found
When creating, updating or deleting a resource which is not
existing, senlinclient desired HTTPNotFound in the current
implementation. But when server requests go to openstacksdk,
openstacksdk will convert all notfound exceptions into its
own exception type ResourceNotFound. So Senlin client should
handle this exception instead.
There is an exception when listing nodes of a cluster, ResourceNotFound
exception will be raised only when printing the output, so there is
no need to catch it when getting the nodes.

Change-Id: Ib80d749e62ed41ae07734a784d309e47c0d3fcca