Avoid import internal classes from SDK (3)
This is part of a series of changes related to the usage of SDK. In some
test cases, we are importing the resource classes from SDK directly and
even instantiate them. This is bad for unit tests. We should instead
mock those resources for unit tests. When functional tests are added,
the real resource list/dump will be tested instead.

Change-Id: Ie79c742cdc0173ad66975adecf935fb9fec8d209