Update python-senlinclient to use current openstacksdk
senlinclient still uses Profile, which is deprecated and marked for
removal. Update create_connection to be able to work in a post-Profile
world. Amongst other things, this was causing nothing to set the
default_microversion setting.

Also, make_client was calling create_connection, but it doesn't need to
because openstackclient already has an sdk_connection object that can be
used. Just return it.

Related to https://storyboard.openstack.org/#!/story/2003146

Change-Id: I7bc0bfdecf1b1461714de1da8b56e3e47154726d